requirements.txt
aiodns==3.2.0
    # via turbo-sticks
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via
    #   discord-py
    #   turbo-sticks
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
async-timeout==5.0.1
    # via turbo-sticks
attrs==25.3.0
    # via aiohttp
certifi==2025.4.26
    # via
    #   requests
    #   turbo-sticks
cffi==1.17.1
    # via
    #   cryptography
    #   pycares
    #   turbo-sticks
charset-normalizer==3.4.1
    # via
    #   requests
    #   turbo-sticks
cryptography==44.0.2
    # via turbo-sticks
discord-py==2.5.2
    # via turbo-sticks
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
gevent==25.4.2
    # via turbo-sticks
greenlet==3.2.1
    # via
    #   gevent
    #   turbo-sticks
idna==3.10
    # via
    #   requests
    #   turbo-sticks
    #   yarl
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   turbo-sticks
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psycopg==3.2.6
    # via turbo-sticks
psycopg-c==3.2.6
    # via psycopg
psycopg-pool==3.2.6
    # via turbo-sticks
pycares==4.6.1
    # via aiodns
pycparser==2.22
    # via cffi
pydantic==2.11.3
    # via
    #   pydantic-settings
    #   turbo-sticks
pydantic-core==2.33.1
    # via
    #   pydantic
    #   turbo-sticks
pydantic-settings==2.9.1
    # via turbo-sticks
python-dotenv==1.1.0
    # via pydantic-settings
pyyaml==6.0.2
    # via turbo-sticks
redis==5.2.1
    # via turbo-sticks
requests==2.32.3
    # via
    #   requests-oauthlib
    #   turbo-sticks
requests-oauthlib==2.0.0
    # via turbo-sticks
.
typing-extensions==4.13.2
    # via
    #   psycopg
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   turbo-sticks
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.4.0
    # via
    #   requests
    #   turbo-sticks
uwsgi==2.0.29
    # via turbo-sticks
yarl==1.20.0
    # via aiohttp
zope-event==5.0
    # via
    #   gevent
    #   turbo-sticks
zope-interface==7.2
    # via
    #   gevent
    #   turbo-sticks

# The following packages were excluded from the output:
# setuptools
